+++ src/services/dragonchain-client/DragonchainClient.ts
@@ -57,6 +57,8 @@ import {
   CustomTagFieldOptions,
   SmartContractLogs,
   PermissionsDocument,
+  L5BlockAtRest,
+  EternalReportV1,
 } from '../../interfaces/DragonchainClientInterfaces';
 import { CredentialService, HmacAlgorithm } from '../credential-service/CredentialService';
 import { getDragonchainId, getDragonchainEndpoint } from '../config-service';
@@ -1435,6 +1437,50 @@ export class DragonchainClient {
     if (options.gasPrice) body.transaction.gasPrice = options.gasPrice;
     if (options.gas) body.transaction.gas = options.gas;
     return (await this.post('/v1/public-blockchain-transaction', body)) as Response<PublicBlockchainTransactionResponse>;
+  };
+
+  /**
+   * Query the Dragonchain for the subsequent interchain (L5) transactions
+   */
+  public queryInterchainTransactions = async (options: {
+    /**
+     * L1 block ID
+     */
+    blockId: string;
+  }) => {
+    if (!options.blockId) throw new FailureByDesign('PARAM_ERROR', 'Parameter `blockId` is required');
+    return (await this.get(`/v1/verifications/interchains/${options.blockId}`)) as Response<L5BlockAtRest[]>;
+  };
+
+  /**
+   * Get/Generate an Eternal-type report given a transaction ID
+   */
+  public getReport = async (options: {
+    /**
+     * the transaction ID of the transaction to generate report for
+     */
+    transactionId: string;
+  }) => {
+    if (!options.transactionId) throw new FailureByDesign('PARAM_ERROR', 'Parameter `transactionId` is required');
+    const transaction = await this.getTransaction({ transactionId: options.transactionId });
+    if (transaction && !transaction.ok) return { ok: false, status: transaction.status, response: {} };
+    const blockId = transaction.response.header.block_id;
+    const block = await this.getBlock({ blockId });
+    if (block && !block.ok) return { ok: false, status: block.status, response: {} };
+    const verifications = await this.getVerifications({ blockId });
+    const l5verifications = await this.queryInterchainTransactions({ blockId });
+    return {
+      ok: true,
+      status: 200,
+      response: {
+        l1Transaction: transaction.response,
+        l1Block: block.response,
+        l2Verifications: verifications.response && verifications.response['2'],
+        l3Verifications: verifications.response && verifications.response['3'],
+        l4Verifications: verifications.response && verifications.response['4'],
+        l5Verifications: l5verifications && l5verifications.response,
+      },
+    } as Response<EternalReportV1>;
   };
 
   /**